First commited as before_46469_rev3

+++ org.eclipse.jgit/src/org/eclipse/jgit/internal/storage/file/ObjectDirectoryPackParser.java
@@ -50,7 +50,6 @@ import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.RandomAccessFile;
-import java.nio.file.StandardCopyOption;
 import java.security.MessageDigest;
 import java.text.MessageFormat;
 import java.util.Arrays;
@@ -367,11 +366,10 @@ public class ObjectDirectoryPackParser extends PackParser {
 
 	@Override
 	protected void onEndThinPack() throws IOException {
-		final byte[] tailHash = this.tailDigest.digest();
 		final byte[] buf = buffer();
 
 		final MessageDigest origDigest = Constants.newMessageDigest();
-		final MessageDigest tailDigest = Constants.newMessageDigest();
+		final MessageDigest tailDigest2 = Constants.newMessageDigest();
 		final MessageDigest packDigest = Constants.newMessageDigest();
 
 		long origRemaining = origEnd;
@@ -394,15 +392,15 @@ public class ObjectDirectoryPackParser extends PackParser {
 				origDigest.update(buf, 0, origCnt);
 				origRemaining -= origCnt;
 				if (origRemaining == 0)
-					tailDigest.update(buf, origCnt, n - origCnt);
+					tailDigest2.update(buf, origCnt, n - origCnt);
 			} else
-				tailDigest.update(buf, 0, n);
+				tailDigest2.update(buf, 0, n);
 
 			packDigest.update(buf, 0, n);
 		}
 
-		if (!Arrays.equals(origDigest.digest(), origHash)
-				|| !Arrays.equals(tailDigest.digest(), tailHash))
+		if (!Arrays.equals(origDigest.digest(), origHash) || !Arrays
+				.equals(tailDigest2.digest(), this.tailDigest.digest()))
 			throw new IOException(
 					JGitText.get().packCorruptedWhileWritingToFilesystem);
 
@@ -478,25 +476,20 @@ public class ObjectDirectoryPackParser extends PackParser {
 			}
 		}
 
-		try {
-			FileUtils.rename(tmpPack, finalPack,
-					StandardCopyOption.ATOMIC_MOVE);
-		} catch (IOException e) {
+		if (!tmpPack.renameTo(finalPack)) {
 			cleanupTemporaryFiles();
 			keep.unlock();
 			throw new IOException(MessageFormat.format(
-					JGitText.get().cannotMovePackTo, finalPack), e);
+					JGitText.get().cannotMovePackTo, finalPack));
 		}
 
-		try {
-			FileUtils.rename(tmpIdx, finalIdx, StandardCopyOption.ATOMIC_MOVE);
-		} catch (IOException e) {
+		if (!tmpIdx.renameTo(finalIdx)) {
 			cleanupTemporaryFiles();
 			keep.unlock();
 			if (!finalPack.delete())
 				finalPack.deleteOnExit();
 			throw new IOException(MessageFormat.format(
-					JGitText.get().cannotMoveIndexTo, finalIdx), e);
+					JGitText.get().cannotMoveIndexTo, finalIdx));
 		}
 
 		try {
